Updated to 3.8.1

+++ crc.hpp
@@ -4,7 +4,7 @@
 extern uint CRCTab[256];
 
 void InitCRC();
-uint CRC(uint StartCRC,const void *Addr,uint Size);
-ushort OldCRC(ushort StartCRC,const void *Addr,uint Size);
+uint CRC(uint StartCRC,const void *Addr,size_t Size);
+ushort OldCRC(ushort StartCRC,const void *Addr,size_t Size);
 
 #endif
